refactor: move toestand from solara.lab to solara
Reactive variables are not experimental anymore and this causes
circular imports for components in solara.lab that use reactive
for typing.
